requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
pytz==2018.9
pygal==2.4.0
django-cleanup==3.2.0
djangorestframework==3.9.3
Pillow==6.0.0
django-jet==1.0.8
feedparser==5.2.1
django-filter==2.1.0
whitenoise==4.1.2
